commit some fixes reclamation

+++ src/main/java/tn/esprit/gestionbancaire/model/Account.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.Date;
 import java.util.List;
@@ -23,7 +24,7 @@ public abstract class Account implements Serializable {
 
     private Date creationDate;
 
-    private float balance;
+    private BigDecimal balance;
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
